Move slow stateful test from cover to nocover
